draft of inteval check edits
